fix rating star

+++ database/migrations/2024_09_25_065218_create_tbl_review.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->increments('review_id');
             $table->integer('user_id');
             $table->integer('product_id');
-            $table->integer('rating');
+            $table->double('rating');
             $table->longText('comment');
             $table->timestamps();
         });
